requirements.txt
black==24.8.0
fastapi==0.115.0
httpx==0.27.2
pydantic-settings==2.5.2
pytest-asyncio==0.24.0
pytest-cov==5.0.0
python-dotenv==1.0.1
twilio==9.3.1
uvicorn==0.30.6